Refactor to move more specific logic
